+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
+## [1.3.0] - 2024-01-03
+
+### Added
+
+- Command: `/lootifications prices` to toggle loot prices in notifications
+- Command: `/lootifications delay reset` to reset the delay time before notifications fade out
+- Command: `/lootifications max reset` to reset the maximum number of displayed notifications
+
 ## [1.2.0] - 2023-12-31
 
-## Added
+### Added
 
 - Command: `/lootifications anchor` to toggle the notifications anchor
 - Command: `/lootifications anchor reset` to reset the notifications anchor